arbitragem_cripto/requirements.txt
# Framework Web
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-socketio==5.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
gevent==22.10.2
gevent-websocket==0.10.1

# Banco de dados
sqlalchemy==2.0.5
sqlalchemy-utils==0.40.0
python-dotenv==1.0.0
alembic==1.9.4

# Processamento de dados
pandas==1.5.3
numpy==1.24.2
python-dateutil==2.8.2

# APIs de criptomoedas
ccxt==3.1.42
requests==2.28.2
websocket-client==1.5.1
aiohttp==3.8.4

# Utilidades
pyjwt==2.6.0
python-dateutil==2.8.2
passlib==1.7.4
email-validator==1.3.1
tenacity==8.2.2
apscheduler==3.10.1
loguru==0.6.0

# Processamento assíncrono
celery==5.2.7
redis==4.5.1

# Testes
pytest==7.2.2
pytest-cov==4.1.0
pytest-mock==3.10.0

# Performance
uvicorn==0.20.0
gunicorn==20.1.0